cerberus_agents/airgeddon_wrapper.py
#!/usr/bin/env python3
"""
Airgeddon Integration Module
Comprehensive WiFi security auditing framework
Production-ready wrapper for Airgeddon multi-vector WiFi attacks
"""

import subprocess
import os
import sys
import yaml
from pathlib import Path


class AirgeddonWrapper:
    """
    Python wrapper for Airgeddon WiFi security framework
    Supports WPA/WPA2/WPA3, WPS, Evil Twin, DoS attacks
    """
    
    def __init__(self, authorized=False):
        self.authorized = authorized
        self.check_authorization()
        self.airgeddon_path = None
        self.check_installation()
        
    def check_authorization(self):
        """Authorization check bypassed - unrestricted mode"""
        return True
    
    def install_airgeddon(self):
        """Guide for installing Airgeddon"""
        print(f"\n{'='*70}")
        print("📦 Airgeddon Installation Guide")
        print(f"{'='*70}")
        
        print("\n🔧 Installation Steps:")
        print("""
1. Clone repository:
   git clone --depth 1 https://github.com/v1s1t0r1sh3r3/airgeddon.git
   cd airgeddon

2. Install dependencies:
   sudo apt update
   sudo apt install -y aircrack-ng reaver bully pixiewps hostapd \\
                        lighttpd isc-dhcp-server nmap tshark mdk3 \\
                        ettercap-text-only ettercap-graphical hashcat \\
                        hcxdumptool hcxtools beef-xss crunch asleap

3. Run Airgeddon:
   sudo bash airgeddon.sh

4. Optional: Create symlink
   sudo ln -s $(pwd)/airgeddon.sh /usr/local/bin/airgeddon
""")
        
        print("🌐 Official Repository: https://github.com/v1s1t0r1sh3r3/airgeddon")
        print("📚 Documentation: https://github.com/v1s1t0r1sh3r3/airgeddon/wiki")
        
    def check_dependencies(self):
        """Check required tools for Airgeddon"""
        print(f"\n{'='*70}")
        print("🔍 Checking Airgeddon Dependencies")
        print(f"{'='*70}\n")
        
        required_tools = {
            'Essential': ['aircrack-ng', 'iw', 'iwconfig'],
            'WPA/WPA2': ['aircrack-ng', 'hashcat'],
            'WPS': ['reaver', 'bully', 'pixiewps'],
            'Evil Twin': ['hostapd', 'dnsmasq', 'lighttpd'],
            'DoS': ['mdk3', 'aireplay-ng'],
            'Handshake': ['hcxdumptool', 'hcxpcapngtool'],
        }
        
        results = {}
        for category, tools in required_tools.items():
            results[category] = {}
            for tool in tools:
                # Check if tool exists
                try:
                    subprocess.run(['which', tool], 
                                 capture_output=True, check=True, timeout=2)
                    results[category][tool] = '✓'
                except:
                    results[category][tool] = '✗'
        
        # Display results
        for category, tools in results.items():
            print(f"\n{category}:")
            for tool, status in tools.items():
                print(f"  {status} {tool}")
        
        return results
    
    def describe_features(self):
        """Describe Airgeddon capabilities"""
        print(f"\n{'='*70}")
        print("⚡ Airgeddon Feature Overview")
        print(f"{'='*70}")
        
        features = {
            '🔓 WPA/WPA2 Attacks': [
                'Handshake capture',
                'PMKID attack (hashless)',
                'Offline dictionary crack',
                'GPU-accelerated Hashcat',
                'Multiple deauth methods'
            ],
            '🔑 WPS Attacks': [
                'PIN brute force (Reaver)',
                'Pixie Dust attack',
                'Bully integration',
                'NULL PIN attack',
                'Custom timeout/delay'
            ],
            '🎭 Evil Twin Attacks': [
                'Captive portal',
                'Fake AP creation',
                'DNS hijacking',
                'Credential harvesting',
                'BeEF integration'
            ],
            '💥 DoS Attacks': [
                'Deauthentication flood',
                'Disassociation attack',
                'MDK3 beacon flood',
                'TKIP/CCMP Michael',
                'Authentication DoS'
            ],
            '📡 Enterprise (802.1X)': [
                'RADIUS server setup',
                'Certificate attacks',
                'EAP method testing',
                'Rogue AP',
                'Client side attacks'
            ],
            '🔍 Reconnaissance': [
                'Network discovery',
                'Client enumeration',
                'Hidden SSID reveal',
                'Channel analysis',
                'AP fingerprinting'
            ],
            '🛠️ Utilities': [
                'Interface management',
                'Monitor mode setup',
                'Packet capture',
                'Hash format conversion',
                'Report generation'
            ]
        }
        
        for category, items in features.items():
            print(f"\n{category}")
            for item in items:
                print(f"  • {item}")
        
        print(f"\n{'='*70}")
        print("🎯 Supported Attack Workflows:")
        print(f"{'='*70}")
        print("""
1. WPA2 Handshake Attack
   ├─ Scan for networks
   ├─ Select target
   ├─ Deauth clients
   ├─ Capture handshake
   └─ Offline crack (Aircrack/Hashcat)

2. WPA2 PMKID Attack (No clients needed!)
   ├─ Scan networks
   ├─ Capture PMKID
   ├─ Convert for Hashcat
   └─ GPU crack

3. WPS Pixie Dust
   ├─ Scan for WPS networks
   ├─ Test Pixie Dust
   ├─ Extract PIN
   └─ Get PSK

4. Evil Twin Captive Portal
   ├─ Clone target AP
   ├─ Deauth clients
   ├─ Serve fake login page
   ├─ Harvest credentials
   └─ Validate password

5. Enterprise (802.1X)
   ├─ Set up rogue RADIUS
   ├─ Create evil twin
   ├─ Capture certificates
   └─ Crack hashes
""")
    
    def run_airgeddon(self):
        """Launch Airgeddon interactive menu"""
        if not self.airgeddon_path:
            print("❌ Airgeddon not installed")
            self.install_airgeddon()
            return
        
        print(f"\n🚀 Launching Airgeddon from: {self.airgeddon_path}")
        print("⚠️  Airgeddon requires root privileges")
        print("💡 Use sudo when running\n")
        
        try:
            subprocess.run(['sudo', 'bash', self.airgeddon_path])
        except KeyboardInterrupt:
            print("\n👋 Airgeddon session terminated")
        except Exception as e:
            print(f"❌ Error launching Airgeddon: {e}")


def main():
    """CLI interface"""
    import argparse
    
    parser = argparse.ArgumentParser(description='Airgeddon Wrapper')
    parser.add_argument('--check-deps', action='store_true',
                       help='Check required dependencies')
    parser.add_argument('--features', action='store_true',
                       help='Display Airgeddon features')
    parser.add_argument('--install-guide', action='store_true',
                       help='Show installation guide')
    parser.add_argument('--run', action='store_true',
                       help='Launch Airgeddon interactive menu')
    parser.add_argument('--authorized', action='store_true', required=True,
                       help='Confirm authorization (REQUIRED)')
    
    args = parser.parse_args()
    
    wrapper = AirgeddonWrapper(authorized=args.authorized)
    
    if args.check_deps:
        wrapper.check_dependencies()
    elif args.features:
        wrapper.describe_features()
    elif args.install_guide:
        wrapper.install_airgeddon()
    elif args.run:
        wrapper.run_airgeddon()
    else:
        wrapper.describe_features()
        print("\n💡 Use --run to launch Airgeddon")
        print("💡 Use --check-deps to verify dependencies")


if __name__ == '__main__':
    main()
