requirements.txt
aiodns>=3.0.0
aiofiles>=23.0.0
aiohttp>=3.8.0
bcrypt>=4.0.1
beautifulsoup4>=4.12.0
boto3>=1.26.0
can-isotp>=2.0.0
coercer
cryptography
deepface>=0.0.79
dnspython>=2.3.0
donpapi
exif>=1.6.0
face_recognition>=1.3.0
google-generativeai>=0.3.0
httpx>=0.23.0
imagehash>=4.3.1
impacket
kube-hunter>=0.6.0
ldap3>=2.9.0
lsassy
paramiko
phonenumbers
Pillow>=10.0.0
playwright>=1.40.0
psycopg2-binary>=2.9.0
publicsuffixlist
pwntools
PyInstaller>=5.13.0
pyinstaller-hooks-contrib>=2023.10
pyjwt>=2.8.0
pymetasploit3>=1.0.0
pymongo>=4.3.0
pymysql>=1.0.0
pytest
pytest-asyncio
python-can>=4.2.0
python-nmap>=0.7.1
python-whois>=0.8.0
pyyaml>=6.0
reportlab
requests>=2.31.0
scapy>=2.5.0
selenium>=4.0.0
shodan>=1.28.0
urllib3>=2.0.0
wapiti3>=3.1.0
